requirements/base.txt
Django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-allauth==0.13.0
django-rest-auth==0.9.5
django-phonenumber-field==7.0.2
pillow==8.4.0
djangorestframework-simplejwt==5.0.0
stripe==2.65.0
drf-yasg==1.21.5